test: move SIPI utils to E2ESpec
